# Lab2/pzpi-22-3-shut-oleksandr-lab2/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users": {
      "get": {
        "operationId": "UsersController_getAllUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all users ",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "User creation",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "Update user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "Get user by id",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "Delete user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/find-by-username/{username}": {
      "get": {
        "operationId": "UsersController_getUserByUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "Get user by username",
        "tags": [
          "users"
        ]
      }
    },
    "/api/roles": {
      "get": {
        "operationId": "RolesController_getRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all roles ",
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "RolesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create role",
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update role",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/roles/{value}": {
      "get": {
        "operationId": "RolesController_getByValue",
        "parameters": [
          {
            "name": "value",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get role by value",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/roles/update-user-role": {
      "patch": {
        "operationId": "RolesController_updateUserRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user role",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/roles/{id}": {
      "delete": {
        "operationId": "RolesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete role",
        "tags": [
          "roles"
        ]
      }
    },
...
    "/api/rooms": {
      "get": {
        "operationId": "RoomsController_getAllRooms",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all rooms",
        "tags": [
          "rooms"
        ]
      },
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Create room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/rooms/{id}/analyze": {
      "get": {
        "operationId": "RoomsController_analyzeRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personalized recommendations"
          }
        },
        "summary": "Analyze room data and generate recommendations",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/rooms/{id}": {
      "get": {
        "operationId": "RoomsController_getRoomById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Get room by id",
        "tags": [
          "rooms"
        ]
      },
      "delete": {
        "operationId": "RoomsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/rooms/update": {
      "post": {
        "operationId": "RoomsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Update room",
        "tags": [
          "rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "Project api",
    "description": "Here stores api documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "cdidk123",
            "description": "username"
          },
          "password": {
            "type": "string",
            "example": "VERY_STRONG_PASSWORD",
            "description": "user password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "12134512df2144",
            "description": "Id of role that u wanna update"
          },
          "username": {
            "type": "string",
            "example": "cdidk123",
            "description": "username"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {}
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "example": "USER",
            "description": "Here you can create role whatever you want"
          },
          "description": {
            "type": "string",
            "example": "default user role",
            "description": "Here you can create role whatever you want"
          }
        },
        "required": [
          "value",
          "description"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "12134512df2144",
            "description": "Id of role that u wanna update"
          },
          "value": {
            "type": "string",
            "example": "name of role actually",
            "description": "It`s about of the role name"
          },
          "description": {
            "type": "string",
            "example": "description of role",
            "description": "It`s about of the role description"
          }
        },
        "required": [
          "id",
          "value",
          "description"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "example": "12134512df2144",
            "description": "Id of user that u wanna update"
          },
          "role_name": {
            "type": "string",
            "example": "USER",
            "description": "name of role"
          }
        },
        "required": [
          "user_id",
          "role_name"
        ]
      },
...
      "Room": {
        "type": "object",
        "properties": {}
      },
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "userId",
            "description": "User ID"
          },
          "temperature": {
            "type": "number",
            "example": 22,
            "description": "Temperature"
          },
          "moisture": {
            "type": "number",
            "example": 50,
            "description": "Moisture"
          },
          "carbonDioxide": {
            "type": "number",
            "example": 400,
            "description": "Carbon Dioxide (CO2)"
          },
          "illumination": {
            "type": "number",
            "example": 300,
            "description": "Illumination"
          }
        },
        "required": [
          "user",
          "temperature",
          "moisture",
          "carbonDioxide",
          "illumination"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "roomID",
            "description": "Room ID"
          },
          "user": {
            "type": "string",
            "example": "userId",
            "description": "User ID"
          },
          "temperature": {
            "type": "number",
            "example": 22,
            "description": "Temperature"
          },
          "moisture": {
            "type": "number",
            "example": 50,
            "description": "Moisture"
          },
          "carbonDioxide": {
            "type": "number",
            "example": 400,
            "description": "Carbon Dioxide (CO2)"
          },
          "illumination": {
            "type": "number",
            "example": 300,
            "description": "Illumination"
          }
        },
        "required": [
          "id",
          "user",
          "temperature",
          "moisture",
          "carbonDioxide",
          "illumination"
        ]
      },
...
    }
  }
}